requirements.txt
flask
feedparser
beautifulsoup4
python-dateutil